updated to start a small static factory and try to get the program class doing the correct thing

+++ PizzaFactory.DAL/Config.cs
@@ -6,7 +6,7 @@ namespace PizzaFactory.DAL
     {
         public double BaseCookingTime { get; set; }
         public double CookingInterval { get; set; }
-        public double RequiredPizzas { get; set; }
+        public int RequiredPizzas { get; set; }
         public double ToppingCookingTime { get; set; }
         public List<DTO.Topping> Toppings { get; set; }
     }
